requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
ansicon==1.89.0
    # via jinxed
appdirs==1.4.4
    # via requests-cache
arrow==1.2.2
    # via django-q
asgiref==3.5.2
    # via django
attrs==21.4.0
    # via
    #   cattrs
    #   requests-cache
black==22.6.0
    # via -r requirements.in
blessed==1.19.1
    # via django-q
cattrs==22.1.0
    # via requests-cache
certifi==2022.6.15
    # via
    #   elasticsearch
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via
    #   pdfminer-six
    #   requests
click==8.1.3
    # via black
colorama==0.4.5
    # via
    #   click
    #   tqdm
cryptography==37.0.4
    # via pdfminer-six
dj-database-url==1.0.0
    # via -r requirements.in
django==4.1
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-picklefield
    #   django-q
django-elasticsearch-dsl==7.2.2
    # via -r requirements.in
django-picklefield==3.1
    # via django-q
django-q==1.3.9
    # via -r requirements.in
elasticsearch==7.17.4
    # via elasticsearch-dsl
elasticsearch-dsl==7.4.0
    # via django-elasticsearch-dsl
exceptiongroup==1.0.0rc8
    # via
    #   cattrs
    #   requests-cache
faker==14.1.0
    # via -r requirements.in
idna==3.3
    # via requests
inflect==6.0.0
    # via -r requirements.in
isort==5.10.1
    # via -r requirements.in
jinja2==3.1.2
    # via -r requirements.in
jinxed==1.2.0
    # via blessed
markupsafe==2.1.1
    # via jinja2
mypy-extensions==0.4.3
    # via black
pathspec==0.9.0
    # via black
pdfminer-six==20220524
    # via pdfplumber
pdfplumber==0.7.4
    # via -r requirements.in
pillow==9.2.0
    # via pdfplumber
platformdirs==2.5.2
    # via black
psycopg2-binary==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==1.9.2
    # via inflect
python-dateutil==2.8.2
    # via
    #   arrow
    #   elasticsearch-dsl
    #   faker
python-dotenv==0.20.0
    # via -r requirements.in
redis==3.5.3
    # via django-q
requests==2.28.1
    # via
    #   -r requirements.in
    #   requests-cache
requests-cache==0.9.5
    # via -r requirements.in
six==1.16.0
    # via
    #   blessed
    #   django-elasticsearch-dsl
    #   elasticsearch-dsl
    #   python-dateutil
    #   url-normalize
sqlparse==0.4.2
    # via django
titlecase==2.4
    # via -r requirements.in
tomli==2.0.1
    # via black
tqdm==4.64.0
    # via -r requirements.in
typing-extensions==4.3.0
    # via pydantic
tzdata==2022.2
    # via django
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.12
    # via
    #   elasticsearch
    #   requests
    #   requests-cache
wand==0.6.10
    # via pdfplumber
wcwidth==0.2.5
    # via blessed
